+++ src/Display/SchemaTable/SchemaTable.js
@@ -3,30 +3,26 @@ import TableHeader from '../SchemaTable/TableHeader/TableHeader';
 import TableRows from '../SchemaTable/TableRows/TableRows';
 import DisplayTable from '../SchemaTable/DisplayTable/DisplayTable';
 import { validateSchema } from '../../lib';
-import { getTableDimensions, getLegend, getTableModel, computeTable } from './lib';
+import { getTableDimensions, getTableModel, computeTable } from '../../lib';
 import { getStyle } from './SchemaTable.style';
 
 const SchemaTable = (props) => {
   const { schema, sentenceLetters, isSavedTable } = props;
-  let { symbols, steps } = schema;
-  const { numRows } = getTableDimensions(sentenceLetters.length);
-  const legend = getLegend(sentenceLetters);
-  const tableModel = getTableModel(symbols, numRows, legend);
+  let { symbols } = schema;
+  let tableModel = getTableModel({...props});
+  const { numRows } = getTableDimensions(sentenceLetters.length)
   const style = getStyle(isSavedTable);
-
+  
   let schemaTableHeaders = (<TableHeader symbols={symbols}/>);
   let schemaTable = (<TableRows symbols={symbols} model={tableModel}/>);
   
   if (validateSchema(symbols)) {
-    const mainOpIndex = computeTable({
-      tableModel: tableModel,
-      schema: schema, 
-      numRows: numRows,
-      steps: steps
-    });
+    const { mainOpIndex } = computeTable(schema, tableModel, numRows);
     schemaTableHeaders = (<TableHeader symbols={symbols} mainOpIndex={mainOpIndex}/>);
     schemaTable = (<TableRows model={tableModel} mainOpIndex={mainOpIndex}/>);
   }
+
+  console.log(schema)
 
   return (
     <DisplayTable
