+++ src/Editor/BuilderButtons/lib.js
@@ -1,7 +1,8 @@
-import { getOpCount } from '../../lib';
+import { getOpCount, getTableModel, computeTable, getTableDimensions } from '../../lib';
 
 const getButtonHandlers = (stateObj) => {
   const {
+    sentenceLetters,
     tutorialSteps, 
     setTutorialSteps, 
     schema, 
@@ -29,9 +30,20 @@ const getButtonHandlers = (stateObj) => {
   const saveValidSchema = () => {
     if (isValid) {
       const steps = getOpCount(symbols);
+      const updateSchema = {...schema, steps: steps, collapsed: false};
+      const { numRows } = getTableDimensions(sentenceLetters.length);
+      const tableModel = getTableModel(stateObj);
+      const { mainOpColumn } = computeTable(updateSchema, tableModel, numRows);
+      const tableInfo = {
+        valid: !mainOpColumn.some(el => el === false),
+        invalid: mainOpColumn.some(el => el === false),
+        satisfiable: mainOpColumn.some(el => el === true),
+        unsatisfiable: !mainOpColumn.some(el => el === true)
+      }
+      updateSchema.tableInfo = tableInfo;
       type === 'P' ? 
-        setPremises([...premises, {...schema, steps: steps, collapsed: false}]) :
-        setConclusion({...schema, steps: steps, collapsed: false});
+        setPremises([...premises, updateSchema ]) :
+        setConclusion(updateSchema);
       setTutorialSteps({...tutorialSteps, saveSchema: true, editorOperator: true});
       setSchema({
         symbols: [],
@@ -44,6 +56,7 @@ const getButtonHandlers = (stateObj) => {
     save: saveValidSchema 
   }
 }
+
